fix type in fighter and attack

+++ core/src/game/sniper_monkey/player/fighter/attack/IAttack.java
@@ -3,5 +3,5 @@ package game.sniper_monkey.player.fighter.attack;
 public interface IAttack {
     void performAttack(float attackFactor);
 
-    int getStaminaCost();
+    float getStaminaCost();
 }
